Final code ready for review
In this code I made a function that would generate a random number. Once I did that, I could select a quote from the JavaScript to make it replace the actual HTML code with the chunk of info that I want from the page. This works to make every refresh generate a random quote in the code.